Address overall review comments: extract shared utilities, hoist imports, fix help command fallback
