+++ lib/resolveDependencies.js
@@ -5,6 +5,7 @@ const path = require("path");
 const os = require("os");
 const semver = require("semver");
 const chalk = require('chalk');
+const { resolveNpmAlias } = require("./utils");
 const CACHE_DIR = path.join(os.homedir(), ".blaze_cache");
 const isVerbose = process.argv.includes('--verbose');
 async function ensureCacheDir() {
@@ -49,40 +50,8 @@ async function resolveDependencies(dependencies, resolved = {}, parent = null, p
             return;
         }
         // Handle npm alias: alias@npm:real-pkg@range
-        if (typeof versionRange === "string" && versionRange.startsWith("npm:")) {
-            // npm:real-pkg@range
-            const match = versionRange.match(/^npm:([^@]+)@(.+)$/);
-            if (match) {
-                const realPkg = match[1];
-                const realRange = match[2];
-                // Alias cycle detection and depth limiting
-                const MAX_ALIAS_DEPTH = 10;
-                if (!worker._aliasChain)
-                    worker._aliasChain = [];
-                if (worker._aliasChain.includes(name)) {
-                    throw new Error(`Alias cycle detected: ${[...worker._aliasChain, name].join(" -> ")}`);
-                }
-                if (worker._aliasChain.length >= MAX_ALIAS_DEPTH) {
-                    throw new Error(`Alias chain too deep (>${MAX_ALIAS_DEPTH}): ${[...worker._aliasChain, name].join(" -> ")}`);
-                }
-                // Push current alias to the chain
-                worker._aliasChain.push(name);
-                try {
-                    // Recursively resolve the real package
-                    await worker([realPkg, realRange]);
-                }
-                finally {
-                    // Pop after recursion
-                    worker._aliasChain.pop();
-                }
-                // Copy the resolved real package under the alias name
-                resolved[name] = {
-                    ...resolved[realPkg],
-                    _alias: true,
-                    _realPackage: realPkg
-                };
-                return;
-            }
+        if (await resolveNpmAlias(name, versionRange, worker, resolved)) {
+            return;
         }
         const meta = await fetchPackageMeta(name, options);
         // Handle special cases like "latest" tag
